requirements.txt
discord.py==2.5.0
python-dotenv==1.0.1
asyncpg~=0.30.0